requirements.txt
absl-py==1.4.0
alembic==1.12.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Automat==22.10.0
Babel==2.12.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
constantly==15.1.0
contourpy==1.1.0
cryptography==41.0.3
cssselect==1.2.0
cycler==0.11.0
datadog-api-client==2.17.0
debugpy==1.7.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.5
docutils==0.20.1
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.12.3
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-MySQL==1.5.2
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==23.5.26
fonttools==4.42.1
fqdn==1.5.1
gast==0.4.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.58.0
h5py==3.9.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipykernel==6.25.2
ipython==8.15.0
isoduration==20.11.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
keras==2.13.1
Kivy==2.2.1
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.6.0
Kivy-Garden==0.1.5
kiwisolver==1.4.5
libclang==16.0.6
lxml==4.9.3
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.1
mpmath==1.3.0
mysql==0.0.3
mysqlclient==2.2.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
notebook==7.0.3
notebook_shim==0.2.3
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.0.76
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.1
pandas==2.1.0
pandocfilters==1.5.0
parsel==1.8.1
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
Protego==0.3.0
protobuf==4.24.3
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
Pygments==2.16.1
PyMySQL==1.1.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
pypiwin32==223
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytube==15.0.0
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
queuelib==1.6.2
referencing==0.30.2
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.2
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.2
Scrapy==2.10.1
Send2Trash==1.8.2
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.20
sqlparse==0.4.4
stack-data==0.6.2
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow-cpu==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tldextract==3.5.0
torch==2.0.1
tornado==6.3.3
traitlets==5.9.0
Twisted==22.10.0
twisted-iocpsupport==1.0.4
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
w3lib==2.1.2
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug==2.3.7
wrapt==1.15.0
WTForms==3.1.0
zope.interface==6.0